+++ server/src/main/java/org/elasticsearch/plugin/score/ExpertScriptPlugin.java
@@ -5,6 +5,9 @@ import org.apache.logging.log4j.Logger;
 import org.apache.lucene.index.LeafReader;
 import org.apache.lucene.index.LeafReaderContext;
 import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.plugin.score.impl.Horspool;
+import org.elasticsearch.plugin.score.impl.MatchScore;
+import org.elasticsearch.plugin.score.impl.TermScore;
 import org.elasticsearch.plugins.Plugin;
 import org.elasticsearch.plugins.ScriptPlugin;
 import org.elasticsearch.script.ScoreScript;
@@ -41,9 +44,9 @@ public class ExpertScriptPlugin extends Plugin implements ScriptPlugin {
         private static final Map<String, ScoreScript.Factory> factoryLookUp = new HashMap<>();
 
         static {
-            factoryLookUp.put("horspool", HorspoolLeafFactory::new);
-            factoryLookUp.put("match_score", MatchScoreLeafFactory::new);
-            factoryLookUp.put("term_score", TermScoreLeafFactory::new);
+            factoryLookUp.put("horspool", HorspoolFactory::new);
+            factoryLookUp.put("match_score", MatchScoreFactory::new);
+            factoryLookUp.put("term_score", TermScoreFactory::new);
         }
 
         @Override
@@ -70,219 +73,6 @@ public class ExpertScriptPlugin extends Plugin implements ScriptPlugin {
         @Override
         public void close() {
             // optionally close resources
-        }
-
-        private static class HorspoolLeafFactory implements LeafFactory {
-            private final Map<String, Object> params;
-            private final SearchLookup lookup;
-            private final List<String> fields;
-            private final String query;
-
-            private HorspoolLeafFactory(
-                Map<String, Object> params, SearchLookup lookup) {
-                if (params.containsKey("field") == false) {
-                    throw new IllegalArgumentException(
-                        "Missing parameter [field]");
-                }
-                if (params.containsKey("query") == false) {
-                    throw new IllegalArgumentException(
-                        "Missing parameter [query]");
-                }
-
-                this.params = params;
-                this.lookup = lookup;
-                fields = (List<String>) params.get("field");
-                query = params.get("query").toString();
-            }
-
-            @Override
-            public boolean needs_score() {
-                return true;  // Return true if the script needs the score
-            }
-
-            @Override
-            public ScoreScript newInstance(LeafReaderContext context)
-                throws IOException {
-                LeafReader reader = context.reader();
-                SourceLookup source = lookup.getLeafSearchLookup(context).source();
-                LeafDocLookup doc = lookup.getLeafSearchLookup(context).doc();
-                return new ScoreScript(params, lookup, context) {
-                    int currentDocid = -1;
-
-                    @Override
-                    public void setDocument(int docid) {
-                        currentDocid = docid;
-                    }
-
-                    @Override
-                    public double execute() {
-                        //获取原来的评分
-                        double rawScore = this.get_score();
-                        final double[] maxScore = {0.0};
-                        fields.forEach(fieldWeight -> {
-                            String[] split = fieldWeight.split("\\^");
-                            String field = split[0];
-                            double weight = split.length == 2 ? Double.parseDouble(split[1]) : 1;
-                            source.setSegmentAndDocument(context, currentDocid);
-                            String value = "";
-                            if (source.containsKey(field)) {
-                                value = String.valueOf(source.get(field));
-                            } else {
-                                return;
-                            }
-                            double score = Horspool.calHorspoolScoreWrapper(value, query);
-//                            double score=MatchScore.scoreWrapper(value,query);
-                            score = score * weight;
-                            if (score > maxScore[0]) {
-                                maxScore[0] = score;
-                            }
-                        });
-                        return maxScore[0];
-                    }
-                };
-            }
-        }
-    }
-
-    private static class MatchScoreLeafFactory implements LeafFactory {
-        private final Map<String, Object> params;
-        private final SearchLookup lookup;
-        private final List<String> fields;
-        private final MatchScore.MatchsMetaInfo matchsMetaInfo;
-
-        private MatchScoreLeafFactory(
-            Map<String, Object> params, SearchLookup lookup) {
-            if (params.containsKey("field") == false) {
-                throw new IllegalArgumentException(
-                    "Missing parameter [field]");
-            }
-            if (params.containsKey("query") == false) {
-                throw new IllegalArgumentException(
-                    "Missing parameter [query]");
-            }
-
-            this.params = params;
-            this.lookup = lookup;
-            fields = (List<String>) params.get("field");
-            String query = params.get("query").toString();
-            this.matchsMetaInfo = MatchScore.MatchsMetaInfo.parseQuery(query);
-        }
-
-        @Override
-        public boolean needs_score() {
-            return true;  // Return true if the script needs the score
-        }
-
-        @Override
-        public ScoreScript newInstance(LeafReaderContext context)
-            throws IOException {
-            LeafReader reader = context.reader();
-            SourceLookup source = lookup.getLeafSearchLookup(context).source();
-            LeafDocLookup doc = lookup.getLeafSearchLookup(context).doc();
-            return new ScoreScript(params, lookup, context) {
-                int currentDocid = -1;
-
-                @Override
-                public void setDocument(int docid) {
-                    currentDocid = docid;
-                }
-
-                @Override
-                public double execute() {
-                    //获取原来的评分
-                    double rawScore = this.get_score();
-                    final double[] maxScore = {0.0};
-                    fields.forEach(fieldWeight -> {
-                        String[] split = fieldWeight.split("\\^");
-                        String field = split[0];
-                        double weight = split.length == 2 ? Double.parseDouble(split[1]) : 1;
-                        source.setSegmentAndDocument(context, currentDocid);
-                        String value = "";
-                        if (source.containsKey(field)) {
-                            value = String.valueOf(source.get(field));
-                        } else {
-                            return;
-                        }
-                        double score = MatchScore.scoreWrapper(value, matchsMetaInfo);
-                        score = score * weight;
-                        if (score > maxScore[0]) {
-                            maxScore[0] = score;
-                        }
-                    });
-                    return maxScore[0];
-                }
-            };
-        }
-    }
-
-    private static class TermScoreLeafFactory implements LeafFactory {
-        private final Map<String, Object> params;
-        private final SearchLookup lookup;
-        private final List<String> fields;
-        private final TermScore.TermsMetaInfo queryMetaInfo;
-
-        private TermScoreLeafFactory(
-            Map<String, Object> params, SearchLookup lookup) {
-            if (params.containsKey("field") == false) {
-                throw new IllegalArgumentException(
-                    "Missing parameter [field]");
-            }
-            if (params.containsKey("query") == false) {
-                throw new IllegalArgumentException(
-                    "Missing parameter [query]");
-            }
-
-            this.params = params;
-            this.lookup = lookup;
-            fields = (List<String>) params.get("field");
-            String query = params.get("query").toString();
-            this.queryMetaInfo = TermScore.TermsMetaInfo.parseQuery(query);
-        }
-
-        @Override
-        public boolean needs_score() {
-            return true;  // Return true if the script needs the score
-        }
-
-        @Override
-        public ScoreScript newInstance(LeafReaderContext context)
-            throws IOException {
-            LeafReader reader = context.reader();
-            SourceLookup source = lookup.getLeafSearchLookup(context).source();
-            LeafDocLookup doc = lookup.getLeafSearchLookup(context).doc();
-            return new ScoreScript(params, lookup, context) {
-                int currentDocid = -1;
-
-                @Override
-                public void setDocument(int docid) {
-                    currentDocid = docid;
-                }
-
-                @Override
-                public double execute() {
-                    //获取原来的评分
-                    double rawScore = this.get_score();
-                    final double[] maxScore = {0.0};
-                    fields.forEach(fieldWeight -> {
-                        String[] split = fieldWeight.split("\\^");
-                        String field = split[0];
-                        double weight = split.length == 2 ? Double.parseDouble(split[1]) : 1;
-                        source.setSegmentAndDocument(context, currentDocid);
-                        String value = "";
-                        if (source.containsKey(field)) {
-                            value = String.valueOf(source.get(field));
-                        } else {
-                            return;
-                        }
-                        double score = TermScore.scoreWrapper(value, queryMetaInfo);
-                        score = score * weight;
-                        if (score > maxScore[0]) {
-                            maxScore[0] = score;
-                        }
-                    });
-                    return maxScore[0];
-                }
-            };
         }
     }
     // end::expert_engine
